refactor: move network info and image picker services to core package
