--- apps/customer/lib/data/core/storage/storage_service.dart
@@ -1,7 +1,0 @@
-abstract class StorageService {
-  Future<void> save<T>(String key, T value);
-  Future<T?> get<T>(String key);
-  Future<void> remove(String key);
-  Future<void> clear();
-  Future<bool> has(String key);
-}
