+++ packages/core/lib/data/network/network_info.dart
