+++ packages/core/lib/services/image_picker_service.dart
